Killing compiler warning
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@71913 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/dbase/dbase.c
@@ -463,7 +463,7 @@ PHP_FUNCTION(dbase_get_record)
 		/* get the value */
 		str_value = (char *)emalloc(cur_f->db_flen + 1);
 
-		if(cursize <= cur_f->db_flen) {
+		if(cursize <= (unsigned)cur_f->db_flen) {
 			cursize = cur_f->db_flen + 1;
 			fnp = erealloc(fnp, cursize);
 		}
